i) reduce write syscalls
go test -bench=. -cpu 1 -benchtime=20s
benchmark                    old ns/op     new ns/op     delta
BenchmarkClientSmallc1       129295        113599        -12.14%
BenchmarkClientSmallc8       107566        95731         -11.00%
BenchmarkClientSmallc64      115366        102583        -11.08%
BenchmarkClientSmallc512     106814        103891        -2.74%

ii) batch writs from different Goroutines if possible
go test -bench=. -cpu 4 -benchtime=20s
benchmark                      old ns/op     new ns/op     delta
BenchmarkClientSmallc1-4       269609        293830        +8.98%
BenchmarkClientSmallc8-4       49062         39327         -19.84%
BenchmarkClientSmallc64-4      46062         32099         -30.31%
BenchmarkClientSmallc512-4     53862         37108         -31.11%

It is worth further investigation on the slowdown when there are 4 processors and only 1 Goroutine even though it is not a common use case. My speculation is that the io batching leads to more thread switches.