@iamqizhao 2015-03-18
That degradation could be measurement noise. Now it is disappearing from my runs. I am going to setup a dedicated env to revisit it.